ISSUE #1898 - Move EmptyStateInfo component to shared style file
