+++ frontend/routes/components/dialogContainer/components/suggestedTreatmentsDialog/suggestedTreatmentsDialog.component.tsx
@@ -24,7 +24,7 @@ import { isEmpty } from 'lodash';
 import { forOwn, pick, uniq } from 'lodash';
 
 import { LabelButton } from '../../../../viewerGui/components/labelButton/labelButton.styles';
-import { EmptyStateInfo } from '../../../../viewerGui/components/views/views.styles';
+import { EmptyStateInfo } from '../../../components.styles';
 import { CellSelect } from '../../../customTable/components/cellSelect/cellSelect.component';
 import {
 	Container, Label, StyledDialogContent, StyledGrid, StyledList, StyledListItem, StyledTypography,
