+++ frontend/routes/components/messagesList/messagesList.styles.ts
@@ -35,16 +35,6 @@ export const LoaderContainer = styled.div`
 	padding: 10px;
 `;
 
-export const EmptyStateInfo = styled.p`
-	padding: 14px;
-	font-size: 13px;
-	color: ${COLOR.BLACK_60};
-	background-color: ${COLOR.BLACK_6};
-	margin: 25px;
-	border-radius: 6px;
-	text-align: center;
-`;
-
 export const Label = styled(Typography)`
 	&& {
 		color: ${COLOR.BLACK_60};
