+++ frontend/routes/viewerGui/components/bim/bim.styles.ts
@@ -24,16 +24,6 @@ export const Container = styled.div`
 	overflow: auto;
 `;
 
-export const EmptyStateInfo = styled.p`
-	padding: 14px;
-	font-size: 13px;
-	color: ${COLOR.BLACK_60};
-	background-color: ${COLOR.BLACK_6};
-	margin: 25px;
-	border-radius: 6px;
-	text-align: center;
-`;
-
 export const Tabs = styled(TabsComponent)`
 	&& {
 		background: ${COLOR.WHITE_87};
