+++ frontend/routes/viewerGui/components/compare/components/compareClash/compareClash.component.tsx
@@ -19,7 +19,7 @@ import React from 'react';
 
 import { TARGET_MODEL_TYPE } from '../../../../../../constants/compare';
 import { renderWhenTrue } from '../../../../../../helpers/rendering';
-import { EmptyStateInfo } from '../../../views/views.styles';
+import { EmptyStateInfo } from '../../../../../components/components.styles';
 import { CompareClashItem } from '../compareClashItem/compareClashItem.component';
 import { CompareFilters } from '../compareFilters/compareFilters.component';
 import { Container, List } from './compareClash.styles';
