+++ frontend/routes/viewerGui/components/views/views.styles.ts
@@ -71,16 +71,6 @@ export const ViewpointItem = styled(MenuItem)`
 	}
 ` as any;
 
-export const EmptyStateInfo = styled.p`
-	padding: 14px;
-	font-size: 13px;
-	color: ${COLOR.BLACK_60};
-	background-color: ${COLOR.BLACK_6};
-	margin: 25px;
-	border-radius: 6px;
-	text-align: center;
-`;
-
 export const NewItemWrapper = styled.div`
 	display: flex;
 	flex: 1;
